unite slyles to parents component
