--- src/components/AuthNavigate/AuthNavigate.jsx
@@ -1,5 +1,0 @@
-import { NavigateButtonStyled } from './NavigateButtonStyled'
-
-export const AuthNavigate = ({ text, route }) => {
-  return <NavigateButtonStyled to={route}>{text}</NavigateButtonStyled>
-}
