+++ src/pages/auth/components/RegisterForm/RegisterForm.styled.js
@@ -1,15 +1,15 @@
 import styled, { css } from 'styled-components'
 import { Field, Form } from 'formik'
-import { getBreakpointsStyles } from '../../styles/breakpoints'
+import { getBreakpointsStyles } from '../../../../styles/breakpoints'
 
 export const Title = styled.h1`
-  ${({ theme: { colors } }) => css`
+  ${({ theme: { colors, shadows } }) => css`
     margin-bottom: 32px;
     font-weight: 600;
     font-size: 18px;
     line-height: 24px;
     color: ${colors.primary};
-    text-shadow: 0px 47px 355px rgba(0, 0, 0, 0.07), 0px 9.4px 57.6875px rgba(0, 0, 0, 0.035);
+    text-shadow: ${shadows.titleShadow};
 
     ${getBreakpointsStyles({
       desktop: css`
@@ -47,18 +47,18 @@ export const FormStyled = styled(Form)`
   `}
 `
 
-export const Input = styled(Field)`
+export const InputField = styled(Field)`
   ${({ theme: { colors } }) => css`
     margin-bottom: 24px;
     padding: 14px;
     width: 287px;
     height: 46px;
     border-radius: 8px;
-    border: 1px solid rgba(220, 227, 229, 0.6);
+    border: 1px solid ${colors.border};
     font-weight: 400;
     font-size: 14px;
     line-height: 18px;
-    color: ${colors.background};
+    color: ${colors.text};
 
     ${getBreakpointsStyles({
       desktop: css`
@@ -79,12 +79,14 @@ export const Input = styled(Field)`
   `}
 `
 export const ErrorMessageStyled = styled.div`
-  position: absolute;
-  bottom: 0;
-  color: #da1414;
-  font-size: 12px;
-  line-height: 14px;
-  margin-left: 8px;
+  ${({ theme: { colors } }) => css`
+    position: absolute;
+    bottom: 0;
+    color: ${colors.textRed};
+    font-size: 12px;
+    line-height: 14px;
+    margin-left: 8px;
+  `}
 `
 
 export const NameOfFieldStyled = styled.div`
