+++ src/styles/theme/themes.js
@@ -7,6 +7,7 @@ const shadows = {
   modalShadow: '0px 4px 16px rgba(17, 17, 17, 0.1)',
   buttonShadow: '4px 2px 16px rgba(136, 165, 191, 0.48)',
   secondButtonShadow: '4px 2px 16px rgba(97, 97, 97, 0.1)',
+  titleShadow: '0px 47px 355px rgba(0, 0, 0, 0.07), 0px 9.4px 57.6875px rgba(0, 0, 0, 0.035)',
 }
 
 export const lightTheme = themeCreator({
@@ -16,6 +17,8 @@ export const lightTheme = themeCreator({
     text: '#111111',
     primary: '#3e85f3',
     hoverPrimary: '#2b78ef',
+    textRed: '#da1414',
+    border: 'rgba(220, 227, 229, 0.6)',
 
     inputBorderDefault: '#DBDBDB',
     placeholderColor: '#DCE3E5',
